separated regSlice
