+++ src/redux/store.js
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
 import userSlice from './slices/userSlice';
+import registerSlice from './slices/registerSlice';
 
 const store = configureStore({
     reducer: {
         user: userSlice,
+        reg: registerSlice,
     },
 });
